refactor(security-hub): extract row components from SecuritySafesTable
Pull the two row variants and the table chrome out of the main file:

- SingleSafeRow.tsx: single-chain Safe row.
- MultichainSafeRow.tsx: collapsed parent row + per-chain children rendered
  when expanded (child row lives as a sub-component in the same file since it
  is tightly coupled to the parent).
- constants.tsx: add TABLE_SX and the COLUMNS list so the table chrome config
  sits beside the other visual constants.

SecuritySafesTable.tsx is now 148 lines (was 684 originally) and just owns
the feature handle, animation/expand state, grade filter, and the
single-vs-multichain dispatch. Storybook snapshot tests pass unchanged.
