+++ apps/web/src/features/spaces/components/SecurityHub/components/SecuritySafesTable/constants.tsx
@@ -1,5 +1,6 @@
 import { forwardRef } from 'react'
 import { TableRow } from '@mui/material'
+import type { SxProps, Theme } from '@mui/material'
 import { motion } from 'framer-motion'
 
 /** Em-dash glyph used wherever a cell has no value to display. */
@@ -22,3 +23,59 @@ export const MotionTableRow = motion.create(
 )
 
 export const MotionTbody = motion.create('tbody')
+
+/**
+ * Top-level `sx` for the Safes table — column borders, rounded row corners,
+ * hover background, and inline typography overrides. Lives here so the main
+ * file isn't dominated by inline style config.
+ */
+export const TABLE_SX: SxProps<Theme> = {
+  tableLayout: 'fixed',
+  borderCollapse: 'separate',
+  borderSpacing: '0 6px',
+  '& th': {
+    border: 'none',
+    borderBottom: 'none !important',
+    py: 0.5,
+    px: 2.5,
+    fontSize: '0.65rem',
+    fontWeight: 700,
+    textTransform: 'uppercase',
+    letterSpacing: '0.5px',
+    color: 'text.primary',
+    opacity: 0.6,
+  },
+  '& td': {
+    border: 'none',
+    height: 72,
+    py: 0,
+    px: 2.5,
+    backgroundColor: 'background.paper',
+    transition: 'background-color 0.12s',
+    verticalAlign: 'middle',
+    '& .MuiTypography-body2': { fontWeight: 500, fontSize: '0.875rem' },
+    '& .MuiTypography-caption': { fontSize: '0.75rem' },
+  },
+  '& tbody tr:hover td': {
+    backgroundColor: 'success.background',
+  },
+  '& th:first-of-type': { pl: 0 },
+  '& td:first-of-type': { borderTopLeftRadius: 12, borderBottomLeftRadius: 12, pl: 3 },
+  '& td:last-of-type': { borderTopRightRadius: 12, borderBottomRightRadius: 12, pr: 3, overflow: 'hidden' },
+}
+
+/**
+ * Column widths for the Safes table header, in display order. Lives next to
+ * `TABLE_SX` so the column layout config is colocated.
+ */
+export const COLUMNS: { label: string; width: string }[] = [
+  { label: 'Account', width: '22%' },
+  { label: 'Network', width: '11%' },
+  { label: 'Balance', width: '9%' },
+  { label: 'Threshold', width: '9%' },
+  { label: 'Version', width: '8%' },
+  { label: 'Status', width: '11%' },
+  { label: 'Score', width: '10%' },
+  { label: 'Last scanned', width: '12%' },
+  { label: '', width: '8%' },
+]
